Code cleanups and Data Helper organization
