+++ Projects/CoX/Common/NetStructures/Friend.cpp
@@ -12,14 +12,18 @@
 
 #include "Friend.h"
 
+#include "Servers/MapServer/DataHelpers.h"
 #include "CharacterHelpers.h"
 #include "GameData/playerdata_definitions.h"
+#include "GameData/map_definitions.h"
 #include "Entity.h"
 #include "Logging.h"
 #include "Character.h"
 
 
-
+/*
+ * Friend Methods -- Sidekick system requires teaming.
+ */
 void toggleFriendList(Entity &src)
 {
     GUIWindow *friendlist = &src.m_player->m_gui.m_wnds.at(WindowIDX::wdw_Friends);
@@ -58,6 +62,88 @@ void dumpFriendsList(const Friend &f)
              << "\n\t" << "origin_id:" << f.m_origin_idx
              << "\n\t" << "map_idx:" << f.m_map_idx
              << "\n\t" << "mapname:" << f.m_mapname;
+}
+
+void addFriend(Entity &src, const Entity &tgt)
+{
+    QString msg;
+    FriendsList &src_data(src.m_char->m_char_data.m_friendlist);
+
+    if(src_data.m_friends_count >= g_max_friends)
+    {
+        msg = "You cannot have more than " + QString::number(g_max_friends) + " friends.";
+        qCDebug(logFriends).noquote() << msg;
+        messageOutput(MessageChannel::USER_ERROR, msg, src);
+        return; // break early
+    }
+
+    src_data.m_has_friends = true;
+    src_data.m_friends_count++;
+
+    Friend f;
+    f.m_online_status   = (tgt.m_client != nullptr); // need some other method for this.
+    f.m_db_id           = tgt.m_db_id;
+    f.m_name            = tgt.name();
+    f.m_class_idx       = tgt.m_entity_data.m_class_idx;
+    f.m_origin_idx      = tgt.m_entity_data.m_origin_idx;
+    f.m_map_idx         = tgt.m_entity_data.m_map_idx;
+    f.m_mapname         = getEntityDisplayMapName(tgt.m_entity_data);
+
+    // add to friendlist
+    src_data.m_friends.emplace_back(f);
+    qCDebug(logFriends) << "friendslist size:" << src_data.m_friends_count << src_data.m_friends.size();
+
+    msg = "Adding " + tgt.name() + " to your friendlist.";
+    qCDebug(logFriends).noquote() << msg;
+    messageOutput(MessageChannel::FRIENDS, msg, src);
+
+    if(logFriends().isDebugEnabled())
+        dumpFriends(src);
+
+    sendFriendsListUpdate(&src, src_data); // Send FriendsListUpdate
+}
+
+void removeFriend(Entity &src, QString friend_name)
+{
+    QString msg;
+    FriendsList &src_data(src.m_char->m_char_data.m_friendlist);
+
+    qCDebug(logFriends) << "Searching for friend" << friend_name << "to remove them.";
+
+    QString lower_name = friend_name.toLower();
+    auto iter = std::find_if( src_data.m_friends.begin(), src_data.m_friends.end(),
+                              [lower_name](const Friend& f)->bool {return lower_name==f.m_name.toLower();});
+
+    if(iter!=src_data.m_friends.end())
+    {
+        msg = "Removing " + iter->m_name + " from your friends list.";
+        iter = src_data.m_friends.erase(iter);
+
+        qCDebug(logFriends) << msg;
+        if(logFriends().isDebugEnabled())
+            dumpFriends(src);
+    }
+    else
+        msg = friend_name + " is not on your friends list.";
+
+    if(src_data.m_friends.empty())
+        src_data.m_has_friends = false;
+
+    src_data.m_friends_count = src_data.m_friends.size();
+
+    qCDebug(logFriends).noquote() << msg;
+    messageOutput(MessageChannel::FRIENDS, msg, src);
+
+    // Send FriendsListUpdate
+    sendFriendsListUpdate(&src, src_data);
+}
+
+const QString &getFriendDisplayMapName(const Friend &f)
+{
+    static const QString offline(QStringLiteral("OFFLINE"));
+    if (!f.m_online_status)
+        return offline;
+    return getDisplayMapName(f.m_map_idx);
 }
 
 //! @}
