+++ Projects/CoX/Common/NetStructures/Friend.h
@@ -32,6 +32,14 @@ static const constexpr  uint32_t    class_version   = 1;
                         std::vector<Friend> m_friends;
 };
 
+
+/*
+ * Friend Methods
+ */
 void toggleFriendList(Entity &src);
 void dumpFriends(const Entity &src);
 void dumpFriendsList(const Friend &f);
+void addFriend(Entity &src, const Entity &tgt);
+void removeFriend(Entity &src, QString friendName);
+
+const QString &getFriendDisplayMapName(const Friend &f);
