+++ Projects/CoX/Common/NetStructures/Team.h
@@ -49,5 +49,15 @@ bool sameTeam(Entity &src, Entity &tgt);
 bool makeTeamLeader(Entity &src, Entity &tgt);
 bool inviteTeam(Entity &src, Entity &tgt);
 bool kickTeam(Entity &tgt);
+void leaveTeam(Entity &e);
 void removeTeamMember(Team &self,Entity *e);
+void findTeamMember(Entity &tgt);
 
+
+/*
+ * Sidekick Methods -- Sidekick system requires teaming.
+ */
+bool isSidekickMentor(const Entity &e);
+void inviteSidekick(Entity &src, Entity &tgt);
+void addSidekick(Entity &tgt, Entity &src);
+void removeSidekick(Entity &src);
