+++ Projects/CoX/Servers/MapServer/Events/GameCommandList.h
@@ -26,6 +26,7 @@
 #include "Events/PowerSystemEvents.h"
 #include "Events/SetClientState.h"
 #include "Events/SidekickOffer.h"
+#include "Events/StandardDialogCmd.h"
 #include "Events/TeamLooking.h"
 #include "Events/TeamOffer.h"
 #include "Events/TrayAdd.h"
